requirements.txt
Flask==3.1.0
gunicorn
azure-monitor-opentelemetry